refactor: improve code base to a better pattern
