+++ app/config/i18n.ts
@@ -2,9 +2,9 @@ import i18next from 'i18next';
 import {initReactI18next} from 'react-i18next';
 import * as RNLocalize from 'react-native-localize';
 
-import {en} from '@/resources/strings/en';
-import {es} from '@/resources/strings/es';
-import {pt} from '@/resources/strings/pt';
+import {en} from '@resources/strings/en';
+import {es} from '@resources/strings/es';
+import {pt} from '@resources/strings/pt';
 
 const resources = {en, es, pt};
